# backend/biosim_server/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: biosim-server
  version: 0.10.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - Users
      summary: Get the authenticated user's profile
      operationId: get-current-user
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - Users
      summary: Delete the authenticated user's account
      operationId: delete-current-user
      responses:
        '204':
          description: Successful Response
        '403':
          description: The token's issuer is not the configured Auth0 tenant, so its
            subject must not be resolved against that tenant's directory. No account
            is deleted.
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Users
      summary: Update the authenticated user's profile
      operationId: update-current-user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '403':
          description: The token's issuer is not the configured Auth0 tenant, so its
            subject must not be resolved against that tenant's directory. No Management
            API call is made.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/password-reset:
    post:
      tags:
      - Users
      summary: Create an Auth0-hosted password reset URL for the authenticated user
      description: 'Issues a short-lived (600 s), single-use Auth0-hosted password-change
        URL for the authenticated primary database user. The account is selected from
        the verified bearer token only -- no email, user id, client id or return URL
        is accepted from the caller, and supplying them changes nothing. No email
        is sent: navigate the browser to the returned URL (it is a bearer capability,
        so it must never be logged, cached, or fetched by script). Issuance is a single
        attempt and is deliberately never retried, because an uncertain outcome may
        already have minted a ticket. Every response, errors included, is `Cache-Control:
        no-store` with `Referrer-Policy: no-referrer`. A 200 means a ticket was issued,
        not that the password was changed.'
      operationId: create-current-user-password-reset
      responses:
        '200':
          description: 'A hosted password-change URL. Sensitive: no-store/no-referrer,
            never persisted by the API, and not proof of a completed change.'
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
            Referrer-Policy:
              description: Always `no-referrer`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
        '401':
          description: Missing, malformed, expired, or otherwise unverifiable bearer
            token. An access token without an expiry is rejected here too.
          headers:
            WWW-Authenticate:
              description: RFC 6750 bearer challenge naming the fault.
              schema:
                type: string
        '403':
          description: 'This principal may not change a password here: its issuer
            is not the configured tenant, its subject is not a primary `auth0|` database
            user, or -- when recent authentication is required -- the token carries
            no usable recent sign-in evidence.'
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
        '429':
          description: The caller exhausted the password-reset budget (metered separately
            from workflow starts, per issuer and subject). Retry after the indicated
            delay; no ticket was issued.
          headers:
            Retry-After:
              description: Seconds until the current fixed window rolls over.
              schema:
                type: integer
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
        '502':
          description: 'Auth0 rejected or failed the ticket request. Generic by design:
            it does not indicate whether the account exists.'
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
        '503':
          description: Password reset is not configured, authentication is temporarily
            unavailable (no usable JWKS), or Auth0 is rate-limiting ticket issuance.
          headers:
            Retry-After:
              description: Present for the Auth0-throttled and JWKS cases.
              schema:
                type: integer
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PasswordResetResponse:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: PasswordResetResponse
...
    UpdateUserProfileRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
      type: object
      title: UpdateUserProfileRequest
    UserProfile:
      properties:
        id:
          type: string
          title: Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        provider:
          type: string
          title: Provider
          description: Auth0 connection derived from the `sub` prefix, e.g. 'google-oauth2'.
        emailVerified:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emailverified
      type: object
      required:
      - id
      - provider
      title: UserProfile
      description: 'API representation of the authenticated user, backing GET/PATCH
        /api/v1/me.


        `id`/`email` come from the validated JWT; `name`/`email_verified` are

        best-effort enrichment from the Auth0 Management API (None when the

        Management API isn''t configured or the call fails).'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer